+++ backend/src/main/java/com/example/backend/common/exception/GlobalExceptionHandler.java
@@ -1,10 +1,8 @@
 package com.example.backend.common.exception;
 
-import com.example.backend.common.enums.ExceptionStatus;
 import com.example.backend.dto.ErrorDto;
 import lombok.extern.slf4j.Slf4j;
 
-import org.springframework.data.crossstore.ChangeSetPersister;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ControllerAdvice;
