+++ backend/src/main/java/com/example/backend/common/exception/NotFoundException.java
@@ -1,8 +1,6 @@
 package com.example.backend.common.exception;
 
-import com.example.backend.common.enums.ExceptionStatus;
 import lombok.Getter;
-import org.springframework.http.HttpStatus;
 
 @Getter
 public class NotFoundException extends RuntimeException {
